uploaded streaming data to s3 bucket

+++ generate_upload_csv.py
@@ -3,6 +3,7 @@ import json
 import os
 import boto3
 from datetime import datetime
+import time
 
 def write_csv_files(json_data, output_directory):
     data = json.loads(json_data)
@@ -36,30 +37,49 @@ def write_csv_files(json_data, output_directory):
             csv_file.close()
 
 
-def upload_csv_files_to_s3(directory_path, bucket_name, s3_prefix='datewise_csv'):
+def upload_csv_files_to_s3(directory_path, s3_prefix):
 
     session = boto3.Session(
     aws_access_key_id='AKIA3JCX6CPUE24DUXOA',
     aws_secret_access_key='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')
     s3 = session.client('s3')
+    bucket_name=''
 
     files = [f for f in os.listdir(directory_path) if f.endswith('.csv')]
 
     for file_name in files:
-        local_file_path = os.path.join(directory_path, file_name)
-        s3_key = os.path.join(s3_prefix, file_name) if s3_prefix else file_name
+        if not file_already_uploaded(file_name):
+            local_file_path = os.path.join(directory_path, file_name)
+            s3_key = os.path.join(s3_prefix, file_name) if s3_prefix else file_name
 
-        try:
-            s3.upload_file(local_file_path, bucket_name, s3_key)
-            s3_url = f"https://{bucket_name}.s3.amazonaws.com/{s3_key}"
-            print(f"Uploaded {local_file_path} to {s3_url}")
-        except Exception as e:
-            print(f"Error uploading {local_file_path} to S3: {str(e)}")
+            try:
+                s3.upload_file(local_file_path, bucket_name, s3_key)
+                s3_url = f"https://{bucket_name}.s3.amazonaws.com/{s3_key}"
+                mark_file_as_uploaded(file_name)
+                print(f"Uploaded {local_file_path} to {s3_url}")
+            except Exception as e:
+                print(f"Error uploading {local_file_path} to S3: {str(e)}")
+
+
+
+def file_already_uploaded(filename):
+    with open('file_logs/file_logs.json','r') as json_file:
+        data=json.load(json_file)
+        if filename in data['file_list']:
+            return True
+
+def mark_file_as_uploaded(filename):
+    with open('file_logs/file_logs.json','r') as json_file:
+        data=json.load(json_file)
+        data['file_list'].append(filename)
+        json_object=json.dumps(data)
+    with open('file_logs/file_logs.json','w') as json_file:
+        json_file.write(json_object)
 
 
 if __name__ == "__main__":
     local_directory = "csv_files"
-    aws_bucket_name = "kafka-s3-connection-bucket"
+    aws_bucket_name = "datalake-stor-poc"
     s3_object_prefix = "sample_csv"
 
     uploaded_urls = upload_csv_files_to_s3(local_directory, aws_bucket_name, s3_object_prefix)
